Updates to example_post_processing.py
Doesn't inject empty lines into code anymore.
Moved testing items to `test_examples_post_processing.py`
Fixed minor typos in documentation.
Correctly strips space and newlines from right end of string before adding newline.
`remove_python_output` should not modify lines above skip.
Find sentences by tesing capital AND now also line length > 12.
Before it just looked for capital, which captured True/False
This is in `remove_python_output`.
In `process_text_block` created add_output function to replace triplicated code.
Also inserted check to avoid printing blank lines in output (happens with large np.arrays)
All read in objects have `\\` replaced with `\\\\`.
There is still an issue with `\` appearing only once in docstrings that needs fixing.
Swapped to using original docstring from codebase, rather than log, for comparing.
Added `new_examples_len` as second measure of increase in code.
